Rename FooException
