+++ src/Exception/File/DuplicateFileException.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace SmartAssert\ApiClient\FooException\File;
+namespace SmartAssert\ApiClient\Exception\File;
 
 class DuplicateFileException extends \Exception
 {
