+++ src/Exception/Http/HttpException.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace SmartAssert\ApiClient\FooException\Http;
+namespace SmartAssert\ApiClient\Exception\Http;
 
 use Psr\Http\Message\RequestInterface;
 use Psr\Http\Message\ResponseInterface;
