+++ src/Exception/Http/UnexpectedContentTypeException.php
@@ -2,17 +2,17 @@
 
 declare(strict_types=1);
 
-namespace SmartAssert\ApiClient\FooException\Http;
+namespace SmartAssert\ApiClient\Exception\Http;
 
 use Psr\Http\Message\RequestInterface;
 use Psr\Http\Message\ResponseInterface;
 
-class UnexpectedDataException extends HttpException
+class UnexpectedContentTypeException extends HttpException
 {
     public function __construct(
         RequestInterface $request,
         ResponseInterface $response,
-        public readonly string $type,
+        public readonly string $contentType,
     ) {
         parent::__construct($request, $response);
     }
