+++ src/Exception/Http/UnexpectedDataException.php
@@ -2,17 +2,17 @@
 
 declare(strict_types=1);
 
-namespace SmartAssert\ApiClient\FooException\Http;
+namespace SmartAssert\ApiClient\Exception\Http;
 
 use Psr\Http\Message\RequestInterface;
 use Psr\Http\Message\ResponseInterface;
 
-class UnexpectedContentTypeException extends HttpException
+class UnexpectedDataException extends HttpException
 {
     public function __construct(
         RequestInterface $request,
         ResponseInterface $response,
-        public readonly string $contentType,
+        public readonly string $type,
     ) {
         parent::__construct($request, $response);
     }
