+++ src/Exception/IncompleteDataException.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace SmartAssert\ApiClient\FooException;
+namespace SmartAssert\ApiClient\Exception;
 
 class IncompleteDataException extends \Exception
 {
