+++ src/Exception/User/AlreadyExistsException.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-namespace SmartAssert\ApiClient\Exception;
+namespace SmartAssert\ApiClient\Exception\User;
 
 use Psr\Http\Message\ResponseInterface;
 
-class UserAlreadyExistsException extends \Exception
+class AlreadyExistsException extends \Exception
 {
     public function __construct(
         public readonly string $email,
