+++ src/FileClient.php
@@ -6,12 +6,12 @@ namespace SmartAssert\ApiClient;
 
 use GuzzleHttp\Psr7\Request as HttpRequest;
 use Psr\Http\Message\ResponseInterface;
-use SmartAssert\ApiClient\FooException\File\DuplicateFileException;
-use SmartAssert\ApiClient\FooException\File\NotFoundException as FileNotFoundException;
-use SmartAssert\ApiClient\FooException\Http\HttpClientException;
-use SmartAssert\ApiClient\FooException\Http\HttpException;
-use SmartAssert\ApiClient\FooException\Http\NotFoundException;
-use SmartAssert\ApiClient\FooException\Http\UnauthorizedException;
+use SmartAssert\ApiClient\Exception\File\DuplicateFileException;
+use SmartAssert\ApiClient\Exception\File\NotFoundException as FileNotFoundException;
+use SmartAssert\ApiClient\Exception\Http\HttpClientException;
+use SmartAssert\ApiClient\Exception\Http\HttpException;
+use SmartAssert\ApiClient\Exception\Http\NotFoundException;
+use SmartAssert\ApiClient\Exception\Http\UnauthorizedException;
 use SmartAssert\ApiClient\ServiceClient\HttpHandler;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
