+++ src/FileSourceClient.php
@@ -6,14 +6,14 @@ namespace SmartAssert\ApiClient;
 
 use GuzzleHttp\Psr7\Request as HttpRequest;
 use SmartAssert\ApiClient\Data\Source\FileSource;
+use SmartAssert\ApiClient\Exception\Http\HttpClientException;
+use SmartAssert\ApiClient\Exception\Http\HttpException;
+use SmartAssert\ApiClient\Exception\Http\NotFoundException;
+use SmartAssert\ApiClient\Exception\Http\UnauthorizedException;
+use SmartAssert\ApiClient\Exception\Http\UnexpectedContentTypeException;
+use SmartAssert\ApiClient\Exception\Http\UnexpectedDataException;
+use SmartAssert\ApiClient\Exception\IncompleteDataException;
 use SmartAssert\ApiClient\Factory\Source\SourceFactory;
-use SmartAssert\ApiClient\FooException\Http\HttpClientException;
-use SmartAssert\ApiClient\FooException\Http\HttpException;
-use SmartAssert\ApiClient\FooException\Http\NotFoundException;
-use SmartAssert\ApiClient\FooException\Http\UnauthorizedException;
-use SmartAssert\ApiClient\FooException\Http\UnexpectedContentTypeException;
-use SmartAssert\ApiClient\FooException\Http\UnexpectedDataException;
-use SmartAssert\ApiClient\FooException\IncompleteDataException;
 use SmartAssert\ApiClient\ServiceClient\HttpHandler;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
