+++ tests/Integration/User/RevokeAllRefreshTokensForUserTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace SmartAssert\ApiClient\Tests\Integration\User;
 
-use SmartAssert\ApiClient\FooException\Http\UnauthorizedException;
+use SmartAssert\ApiClient\Exception\Http\UnauthorizedException;
 use SmartAssert\ApiClient\Tests\Integration\AbstractIntegrationTestCase;
 
 class RevokeAllRefreshTokensForUserTest extends AbstractIntegrationTestCase
